components/MenuBackground: Implement MenuBackgroundComponent
Because of the change into 'ngx-lottie', there is a need to make the background a separate component. We can implement them on the pages, but that will duplicate our code. Since we're going to use the same background Lottie overlay for most of them anyway, do it as a component.

Once we decide that different Lottie animations are used as backgrounds, we can just make the component *very* flexible. Still a win-win.