services/GameData: Encode data being stored and received
Encoded in Base64, this should restrict non-savvy users to decrypt answer keys and edit scores (which isn't possible on Android unless you're rooted, but anyway). Not strong but should be enough for this, I guess.

To maintain performance, currentLevelData for each category is now stored in memory, instead of being encoded and decoded on demand. Loading them wouldn't take a hit anyway because they're hidden behind loading screens.